Move verilator coverage tests into seperate file. Improve VCS coverage option adding
