+++ src/test/scala/chiseltest/tests/OptionsPassingTest.scala
@@ -92,108 +92,4 @@ class OptionsPassingTest extends AnyFlatSpec with ChiselScalatestTester with Mat
     output.contains("Symbol table:") should be(true)
     output.contains("clock/prev") should be(true)
   }
-
-  it should "allow specifying toggle coverage for Verilator" in {
-    val coverageName = "test_run_dir/Testers2_should_allow_specifying_toggle_coverage_for_Verilator/logs/coverage.dat"
-    val coverage = new File(coverageName)
-    val outputStream = new ByteArrayOutputStream()
-    Console.withOut(new PrintStream(outputStream)) {
-      test(new Module {
-        val io = IO(new Bundle {
-          val a = Input(UInt(8.W))
-          val b = Output(UInt(8.W))
-        })
-        io.b := io.a
-      }).withAnnotations(Seq(VerilatorBackendAnnotation, ToggleCoverageAnnotation)) { c =>
-        c.io.a.poke(1.U)
-        c.io.b.expect(1.U)
-        c.clock.step()
-        c.io.a.poke(42.U)
-        c.io.b.expect(42.U)
-      }
-    }
-    val output = outputStream.toString
-    coverage.exists() should be(true)
-    output.contains("--coverage-toggle") should be(true)
-    output.contains("--coverage-line") should be(false)
-    output.contains("--coverage-user") should be(false)
-  }
-
-  it should "allow specifying line coverage for Verilator" in {
-    val coverageName = "test_run_dir/Testers2_should_allow_specifying_line_coverage_for_Verilator/logs/coverage.dat"
-    val coverage = new File(coverageName)
-    val outputStream = new ByteArrayOutputStream()
-    Console.withOut(new PrintStream(outputStream)) {
-      test(new Module {
-        val io = IO(new Bundle {
-          val a = Input(UInt(8.W))
-          val b = Output(UInt(8.W))
-        })
-        io.b := io.a
-      }).withAnnotations(Seq(VerilatorBackendAnnotation, LineCoverageAnnotation)) { c =>
-        c.io.a.poke(1.U)
-        c.io.b.expect(1.U)
-        c.clock.step()
-        c.io.a.poke(42.U)
-        c.io.b.expect(42.U)
-      }
-    }
-    val output = outputStream.toString
-    coverage.exists() should be(true)
-    output.contains("--coverage-toggle") should be(false)
-    output.contains("--coverage-line") should be(true)
-    output.contains("--coverage-user") should be(false)
-  }
-
-  it should "allow specifying structural coverage for Verilator" in {
-    val coverageName = "test_run_dir/Testers2_should_allow_specifying_structural_coverage_for_Verilator/logs/coverage.dat"
-    val coverage = new File(coverageName)
-    val outputStream = new ByteArrayOutputStream()
-    Console.withOut(new PrintStream(outputStream)) {
-      test(new Module {
-        val io = IO(new Bundle {
-          val a = Input(UInt(8.W))
-          val b = Output(UInt(8.W))
-        })
-        io.b := io.a
-      }).withAnnotations(Seq(VerilatorBackendAnnotation, StructuralCoverageAnnotation)) { c =>
-        c.io.a.poke(1.U)
-        c.io.b.expect(1.U)
-        c.clock.step()
-        c.io.a.poke(42.U)
-        c.io.b.expect(42.U)
-      }
-    }
-    val output = outputStream.toString
-    coverage.exists() should be(true)
-    output.contains("--coverage-toggle") should be(true)
-    output.contains("--coverage-line") should be(true)
-    output.contains("--coverage-user") should be(false)
-  }
-
-  it should "allow specifying user coverage for Verilator" in {
-    val coverageName = "test_run_dir/Testers2_should_allow_specifying_user_coverage_for_Verilator/logs/coverage.dat"
-    val coverage = new File(coverageName)
-    val outputStream = new ByteArrayOutputStream()
-    Console.withOut(new PrintStream(outputStream)) {
-      test(new Module {
-        val io = IO(new Bundle {
-          val a = Input(UInt(8.W))
-          val b = Output(UInt(8.W))
-        })
-        io.b := io.a
-      }).withAnnotations(Seq(VerilatorBackendAnnotation, UserCoverageAnnotation)) { c =>
-        c.io.a.poke(1.U)
-        c.io.b.expect(1.U)
-        c.clock.step()
-        c.io.a.poke(42.U)
-        c.io.b.expect(42.U)
-      }
-    }
-    val output = outputStream.toString
-    coverage.exists() should be(true)
-    output.contains("--coverage-toggle") should be(false)
-    output.contains("--coverage-line") should be(false)
-    output.contains("--coverage-user") should be(true)
-  }
 }
